+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Nuke:
 
 * Added support for wildcard searches of mattes.
 * The "Expand Wildcards" knob controls whether the wildcards are expanded or not.
-  * If the wildcards are not expanded, the manifest is recomputed at every frame.
+  * If the wildcards are not expanded, wildcards are not used.
 
 ### 1.2.5